raspi_deployer_starter/requirements.txt
fastapi
uvicorn[standard]
fabric
invoke
paramiko
python-dotenv
jinja2
pytest
requests
zeroconf
psutil
pyjwt
pyyaml
python-multipart
